// sources/TaskPlaner/Web.App/web-app/src/TaskAdministration/Interfaces/ITaskItemProps.ts
import { TaskStatusEnum } from "../../Lib/Enums/TaskStatusEnum";
import { TaskTypeEnum } from "../../Lib/Enums/TaskTypeEnum";

export interface ITaskItemProps {
  id: number;
  title: string;
  status: TaskStatusEnum;
  shortDescription: string;
  assignedUserId?: number;
  taskTypeId?: TaskTypeEnum;
}